# public/api/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "group.one Centralized License Service API",
        "description": "Multi-tenant, multi-brand license management system API.\n\nAll API requests support the `X-Correlation-ID` header for request tracing. If not provided, one will be generated and returned in the response.",
        "contact": {
            "email": "support@licenseservice.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/v1/licenses": {
            "post": {
                "tags": [
                    "Licenses"
                ],
                "summary": "Create a new license",
                "description": "Create a new license for a customer. If no license key exists for the customer email, one will be generated.",
                "operationId": "createLicense",
                "parameters": [
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Optional correlation ID for request tracing",
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_email",
                                    "product_name",
                                    "product_slug",
                                    "license_type"
                                ],
                                "properties": {
                                    "customer_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "customer@example.com"
                                    },
                                    "customer_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "product_name": {
                                        "description": "Human-readable product name",
                                        "type": "string",
                                        "example": "WP Rocket Pro"
                                    },
                                    "product_slug": {
                                        "description": "Machine-friendly product identifier (lowercase, hyphen-separated)",
                                        "type": "string",
                                        "example": "wp-rocket-pro"
                                    },
                                    "license_type": {
                                        "type": "string",
                                        "enum": [
                                            "perpetual",
                                            "subscription",
                                            "trial"
                                        ],
                                        "example": "subscription"
                                    },
                                    "max_activations_per_instance": {
                                        "description": "Maximum seats (unique values) per instance type for seat management. Keys are instance types (site_url, host, machine_id), values are the max number of unique values allowed.",
                                        "type": "object",
                                        "example": {
                                            "site_url": 3,
                                            "host": 5,
                                            "machine_id": 1
                                        }
                                    },
                                    "expires_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "metadata": {
                                        "type": "object",
                                        "example": {
                                            "plan": "pro"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "License created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "license": {
                                                    "$ref": "#/components/schemas/License"
                                                },
                                                "license_key": {
                                                    "description": "Only returned when a new license key is generated",
                                                    "type": "string",
                                                    "example": "ABCDE-FGHIJ-KLMNO-PQRST-UVWXY"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "ApiKeyHeader": []
                    }
                ]
            }
        },
        "/v1/licenses/{id}/renew": {
            "post": {
                "tags": [
                    "Licenses"
                ],
                "summary": "Renew a license",
                "description": "Extend the expiration date of a license. Only applicable to subscription and trial licenses.",
                "operationId": "renewLicense",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "License ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "expires_at"
                                ],
                                "properties": {
                                    "expires_at": {
                                        "description": "New expiration date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "License renewed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "License renewed successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/License"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "License not found"
                    },
                    "422": {
                        "description": "License cannot be renewed (e.g., perpetual license)"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "ApiKeyHeader": []
                    }
                ]
            }
        },
        "/v1/licenses/{id}/suspend": {
            "post": {
                "tags": [
                    "Licenses"
                ],
                "summary": "Suspend a license",
                "description": "Temporarily suspend a license. Suspended licenses cannot be used for activations.",
                "operationId": "suspendLicense",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "License ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "description": "Optional reason for suspension",
                                        "type": "string",
                                        "example": "Payment overdue"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "License suspended successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "License suspended successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/License"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "License not found"
                    },
                    "422": {
                        "description": "License is already suspended or cancelled"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "ApiKeyHeader": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Brand": {
                "title": "Brand",
                "description": "Brand (Tenant) model",
                "required": [
                    "id",
                    "name",
                    "slug"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "name": {
                        "description": "Brand name",
                        "type": "string",
                        "example": "Acme Corporation"
                    },
                    "slug": {
                        "description": "URL-friendly identifier",
                        "type": "string",
                        "example": "acme-corp"
                    },
                    "description": {
                        "type": "string",
                        "example": "Leading software company",
                        "nullable": true
                    },
                    "contact_email": {
                        "type": "string",
                        "format": "email",
                        "example": "contact@acme.com",
                        "nullable": true
                    },
                    "website": {
                        "type": "string",
                        "format": "url",
                        "example": "https://acme.com",
                        "nullable": true
                    },
                    "settings": {
                        "type": "object",
                        "example": {
                            "theme": "dark"
                        },
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "License": {
                "title": "License",
                "description": "License model",
                "required": [
                    "id",
                    "brand_id",
                    "license_key_id",
                    "product_name",
                    "product_slug",
                    "license_type"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "brand_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    },
                    "license_key_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440002"
                    },
                    "product_name": {
                        "description": "Human-readable product name",
                        "type": "string",
                        "example": "WP Rocket Pro"
                    },
                    "product_slug": {
                        "description": "Machine-friendly product identifier",
                        "type": "string",
                        "example": "wp-rocket-pro"
                    },
                    "license_type": {
                        "type": "string",
                        "enum": [
                            "perpetual",
                            "subscription",
                            "trial"
                        ],
                        "example": "subscription"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "suspended",
                            "cancelled",
                            "expired"
                        ],
                        "example": "active"
                    },
                    "max_activations_per_instance": {
                        "description": "Maximum seats (unique values) allowed per instance type for seat management",
                        "type": "object",
                        "example": {
                            "site_url": 3,
                            "host": 5,
                            "machine_id": 1
                        }
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-12-31T23:59:59Z",
                        "nullable": true
                    },
                    "metadata": {
                        "type": "object",
                        "example": {
                            "plan": "pro"
                        },
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "brand": {
                        "$ref": "#/components/schemas/Brand"
                    },
                    "license_key": {
                        "$ref": "#/components/schemas/LicenseKey"
                    }
                },
                "type": "object"
            },
            "LicenseKey": {
                "title": "License Key",
                "description": "License key model - keys are formatted as 5 groups of 5 characters",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "brand_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    },
                    "customer_email": {
                        "type": "string",
                        "format": "email",
                        "example": "customer@mailinator.com"
                    },
                    "key": {
                        "description": "License key in 5 groups of 5 characters",
                        "type": "string",
                        "example": "ABCDE-FGHIJ-KLMNO-PQRST-UVWXY"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "suspended",
                            "cancelled",
                            "expired"
                        ],
                        "example": "active"
                    },
                    "max_activations_per_instance": {
                        "description": "Maximum seats (unique values) allowed per instance type for seat management",
                        "type": "object",
                        "example": {
                            "site_url": 3,
                            "host": 5,
                            "machine_id": 1
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "SanctumAuth": {
                "type": "http",
                "description": "Laravel Sanctum Bearer token obtained from /v1/auth/login",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            },
            "ApiKeyAuth": {
                "type": "http",
                "description": "API key in the format: lcs_xxxxxxxx.xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx",
                "bearerFormat": "API Key",
                "scheme": "bearer"
            },
            "ApiKeyHeader": {
                "type": "apiKey",
                "description": "API key for authentication via header",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
}